FIX: Display confirmation dialog when editing group auto email domains
