+++ app/assets/javascripts/discourse/app/controllers/groups-new.js
@@ -1,42 +1,12 @@
 import Controller from "@ember/controller";
 import { action } from "@ember/object";
 import { service } from "@ember/service";
-import { ajax } from "discourse/lib/ajax";
 import { popupAjaxError } from "discourse/lib/ajax-error";
 import discourseComputed from "discourse/lib/decorators";
-import { i18n } from "discourse-i18n";
-
-export function popupAutomaticMembershipAlert(group_id, email_domains) {
-  if (!email_domains) {
-    return;
-  }
-
-  const data = {};
-  data.automatic_membership_email_domains = email_domains;
-
-  if (group_id) {
-    data.id = group_id;
-  }
-
-  ajax(`/admin/groups/automatic_membership_count.json`, {
-    type: "PUT",
-    data,
-  }).then((result) => {
-    const count = result.user_count;
-
-    if (count > 0) {
-      this.dialog.alert(
-        i18n("admin.groups.manage.membership.automatic_membership_user_count", {
-          count,
-        })
-      );
-    }
-  });
-}
 
 export default class GroupsNewController extends Controller {
-  @service dialog;
   @service router;
+  @service group_automatic_members_dialog;
 
   saving = null;
 
@@ -51,14 +21,19 @@ export default class GroupsNewController extends Controller {
   }
 
   @action
-  save() {
+  async save() {
     this.set("saving", true);
     const group = this.model;
 
-    popupAutomaticMembershipAlert(
+    const accepted = await this.group_automatic_members_dialog.showConfirm(
       group.id,
       group.automatic_membership_email_domains
     );
+
+    if (!accepted) {
+      this.set("saving", false);
+      return;
+    }
 
     group
       .create()
